modificando os dados para jwt/token

+++ src/pessoas/pessoas.service.ts
@@ -1,10 +1,11 @@
-import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreatePessoaDto } from './dto/create-pessoa.dto';
 import { UpdatePessoaDto } from './dto/update-pessoa.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Pessoa } from './entities/pessoa.entity';
 import { Repository } from 'typeorm';
 import { HashingService } from 'src/auth/hashing.service';
+import { TokenPayloadDto } from 'src/auth/DTO/token-payload.dto';
 
 @Injectable()
 export class PessoasService {
@@ -62,7 +63,7 @@ export class PessoasService {
     return pessoa;
   }
 
-  async update(id: number, updatePessoaDto: UpdatePessoaDto) {
+  async update(id: number, updatePessoaDto: UpdatePessoaDto, tokenPayLoad: TokenPayloadDto) {
     const dadosPessoa = {
       nome: updatePessoaDto?.nome,  /// basicamaente esse diacho e a atualizacao do user, sendo opcional e dai ele armazena numa variavel
     };
@@ -85,16 +86,23 @@ export class PessoasService {
       throw new NotFoundException('Pessoa não encontrada')
     }
 
+    if(pessoa.id!== tokenPayLoad.sub){
+      throw new ForbiddenException("Voce nao é essa pessoa");
+    }
+
     return this.pessoaRepository.save(pessoa)
   }
 
-  async remove(id: number) {
+  async remove(id: number, tokenPayLoad: TokenPayloadDto) {
     const pessoa = await this.pessoaRepository.findOneBy({
       id
     })
 
     if (!pessoa) {
       throw new NotFoundException('Pessoa não encontrada')
+    }
+    if(pessoa.id!== tokenPayLoad.sub){
+      throw new ForbiddenException("Voce nao é essa pessoa");
     }
 
     return this.pessoaRepository.remove(pessoa)
